requirements.txt
pyyaml==6.0.1
streamlit==1.49.1
pytest==7.4.4
